moved autoChooser to Dashboard
Hopefully we won't get this error again.
